added a code to clean database when running seed.ts
