+++ client/src/main.tsx
@@ -6,8 +6,8 @@ import App from './App.jsx';
 // import Home from './pages/Home';  UNCOMMENT IF WE END UP USING THIS
 import Signup from './pages/Signup';
 import Login from './pages/Login';
-import SingleThought from './pages/SingleThought';
-import Profile from './pages/Profile';
+// import SingleThought from './pages/SingleThought';
+// import Profile from './pages/Profile';
 import ErrorPage from './pages/Error';
 import Hangman from './pages/Hangman.js';
 
@@ -26,16 +26,18 @@ const router = createBrowserRouter([
       }, {
         path: '/signup',
         element: <Signup />
-      }, {
-        path: '/profiles/:username',
-        element: <Profile />
-      }, {
-        path: '/me',
-        element: <Profile />
-      }, {
-        path: '/thoughts/:thoughtId',
-        element: <SingleThought />
       }, 
+      // {
+      //   path: '/profiles/:username',
+      //   element: <Profile />
+      // }, {
+      //   path: '/me',
+      //   element: <Profile />
+      // }, 
+      // {
+      //   path: '/thoughts/:thoughtId',
+      //   element: <SingleThought />
+      // }, 
     ]
   },
 ]);
